reorganized_framework/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "informatica-spark-framework"
version = "2.0.0"
description = "Framework for converting Informatica BDM projects to PySpark applications"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Informatica Spark Team", email = "info@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "pyspark>=3.2.0",
    "jinja2>=3.0.0", 
    "pyyaml>=6.0",
    "lxml>=4.6.0",
    "packaging>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/informatica/spark-framework"
Documentation = "https://informatica-spark-framework.readthedocs.io"
Repository = "https://github.com/informatica/spark-framework.git"
Issues = "https://github.com/informatica/spark-framework/issues"

[project.scripts]
informatica-spark = "informatica_spark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true